// extensions/xcode/src/models/xcode-project/xcode-project-type.model.ts
/**
 * A XcodeProjectType
 */
export enum XcodeProjectType {
  /**
   * Xcode Project
   */
  project = "xcodeproj",
  /**
   * Xcode Workspace
   */
  workspace = "xcworkspace",
  /**
   * Swift Package
   */
  swiftPackage = "swift",
  /**
   * Swift Playground
   */
  swiftPlayground = "playground",
}